// frontend/src/api/generated/models/journalEntryResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface JournalEntryResponse {
  id: string;
  period_id: string;
  date: string;
  description: string;
  reference: string | null;
  status: string;
}